j.ui: canvastoscreen is expecting an address of a long and we pass an address of an int. on 64-bit the long is 64-bits and the in is 32-bits -- now fixed. see #714.

+++ source/j.ui/j.ui.cpp
@@ -1123,7 +1123,7 @@ void ui_menu_do(t_ui *x, t_object *patcherview, t_pt px, long modifiers)
 	long				size, i;
 	int 				selectedId;
 	t_jfont 			*font;
-	int					coord_x=0, coord_y=0;
+	long				coord_x=0, coord_y=0;
 	t_pt				pt;
 	
 	ui_menu_build(x);	// would be better to not rebuild the menu every single time?  or not?  this uses less memory...
@@ -1305,7 +1305,7 @@ void ui_refmenu_do(t_ui *x, t_object *patcherview, t_pt px, long modifiers)
 	long				size, i;
 	int 				selectedId;
 	t_jfont 			*font;
-	int					coord_x=0, coord_y=0;
+	long				coord_x=0, coord_y=0;
 	t_pt				pt;
 	
 	ui_refmenu_build(x);	// TODO: would be better to not rebuild the menu every single time?  or not?  this uses less memory...
